requirements.txt
asttokens==2.0.5
autopep8==1.5.7
colorama==0.4.4
executing==0.8.2
icecream==2.1.1
mecab-python3==1.0.4
pycodestyle==2.7.0
Pygments==2.10.0
six==1.16.0
toml==0.10.2
